Rename rng.h to randombytes.h and use size_t in argument
Idea borrowed from
MAYO-C.git, commit fc9079fb5ac5cd4af98e3e0f094a0a3cf2a01499.

Makes it easier to integrate SQISign in liboqs. By including
randombytes.h, the pqclean_shims in liboqs.git/src/common/pqclean_shims
will be used.

Only compile-tested on amd64.

+++ src/common/arm64crypto/randombytes_ctrdrbg_inline_asm.c
@@ -174,7 +174,7 @@ void randombytes_init_arm64crypto(unsigned char *entropy_input,
 
 #define WAYS 4
 
-int randombytes_arm64crypto(unsigned char *x, unsigned long long xlen) {
+int randombytes_arm64crypto(unsigned char *x, size_t xlen) {
   uint8_t subkeys[15][16];
   unsigned char block[16];
   u128_t V[WAYS], Vle[WAYS];
@@ -405,7 +405,7 @@ int randombytes_arm64crypto(unsigned char *x, unsigned long long xlen) {
 }
 
 #ifdef RANDOMBYTES_ARM64CRYPTO
-int randombytes(unsigned char *random_array, unsigned long long nbytes) {
+int randombytes(unsigned char *random_array, size_t nbytes) {
   int ret = randombytes_arm64crypto(random_array, nbytes);
 #ifdef ENABLE_CT_TESTING
   VALGRIND_MAKE_MEM_UNDEFINED(random_array, ret);
